refactor: remove V1 content detection bridge, migrate to V2 scoring
Remove the entire V1 content detection bridge that converted V2 additive
scoring (0.0-5.0) back to V1 format (CheckResultType enum + Confidence
0-100). This was losing ProcessingTimeMs data (#311) and leaving the
Algorithm Performance analytics table empty.

Changes:
- Delete dead V1 types: ContentCheckResponse, ContentCheckResult,
  IContentCheck, IContentDetector, ContentCheckHelpers, AIResponseParser,
  CheckResultType enum
- Rename DB columns: confidence→score, net_confidence→net_score (double)
- Migrate all JSONB check_results_json from V1 to V2 format
- Migrate config thresholds from V1 int scale (0-100) to V2 double (0.0-5.0)
- Update all sub-config ConfidenceThreshold properties from int to double
- Freeze view SQL in previous migrations before updating view consts
- Remove ConvertV2ResponsesToV1() from ContentDetectionEngineV2
- Update ContentDetectionResult to use TotalScore instead of V1 fields
- Update all downstream consumers, Razor pages, and tests

Closes #311

+++ TelegramGroupsAdmin.ContentDetection/Models/ImageCheckRequest.cs
@@ -9,5 +9,5 @@ public sealed class ImageCheckRequest : ContentCheckRequestBase
     public required string? PhotoUrl { get; init; }
     public string? PhotoLocalPath { get; init; } // ML-5: Local file path for OCR (optional)
     public required string? CustomPrompt { get; init; }
-    public required int ConfidenceThreshold { get; init; }
+    public required double ConfidenceThreshold { get; init; }
 }
