+++ TelegramGroupsAdmin.ContentDetection/Models/InvisibleCharsCheckRequest.cs
@@ -5,5 +5,5 @@ namespace TelegramGroupsAdmin.ContentDetection.Models;
 /// </summary>
 public sealed class InvisibleCharsCheckRequest : ContentCheckRequestBase
 {
-    public required int ConfidenceThreshold { get; init; }
+    public required double ConfidenceThreshold { get; init; }
 }
