+++ TelegramGroupsAdmin.ContentDetection/Models/VideoCheckRequest.cs
@@ -7,5 +7,5 @@ public sealed class VideoCheckRequest : ContentCheckRequestBase
 {
     public required string VideoLocalPath { get; init; }
     public required string? CustomPrompt { get; init; }
-    public required int ConfidenceThreshold { get; init; }
+    public required double ConfidenceThreshold { get; init; }
 }
